Improve command parsing, add prefix_only param
Currently, morgenbot will respond to !commands anywhere in the string. However, if the command isn't the first thing in the line it doesn't parse the arguments correctly, and if the command is unrecognized, it chops out the wrong thing when assembling the giphy parameter.

This PR fixes the arguments parsing and the giphy string and adds a configuration flag to only recognize !commands at the beginning of the input string, for users that desire that stricter behavior.